ENH: Using the ITK filter container for all input images
Previously, separate containers was used for fixed and moving images.
The ITK pipeline would be driven only by the first fixed and first moving images.
This commit introduces extra bookeeping so that all images can be added to the
ITK input container and split into seperate containers for elastix at a later
stage. This is necessary because only ITK input container is part of the
pipeline mechanism.

Another (and possibly simpler) solution would be override the Update() function
and check that all image containers are up to date. However, this places the
responsilibity of pipeline mechanism maintenance on us. It is probably better
to keep this upstream.